pyproject.toml
[project]
name = "reasoning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Anonymous", email = "Anonymous" }
]
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.8.0",
    "datasets>=3.6.0",
    "flashinfer-python>=0.2.8",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "ipython>=9.3.0",
    "ipywidgets>=8.1.7",
    "jaxtyping>=0.3.2",
    "loguru>=0.7.3",
    "pandas>=2.3.0",
    "pip>=25.1.1",
    "pylatexenc>=2.10",
    "seaborn>=0.13.2",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.47.1",
    "sympy>=1.14.0",
    "tabulate>=0.9.0",
    "torch>=2.7.1",
    "trackio>=0.2.2",
    "transformers>=4.52.4",
    "vllm>=0.10.0",
]

[project.scripts]
reasoning = "reasoning:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["F722"]

[dependency-groups]
dev = [
    "libcst>=1.8.2",
]
